refactor config to "environment"
