+++ packages/functions/src/config/environment.ts
@@ -1,7 +1,7 @@
-import doetenv from 'dotenv';
-doetenv.config();
+import dotenv from 'dotenv';
+dotenv.config();
 
-const config = {
+const environment = {
   supabaseUrl: process.env.SUPABASE_URL ?? '',
   supabasePublicKey: process.env.SUPABASE_KEY ?? '',
   supabasePluginTableName: process.env.SUPABASE_PLUGIN_TABLE_NAME ?? '',
@@ -10,4 +10,4 @@ const config = {
   functionsAlertWebHookUrl: process.env.FUNCTIONS_ALERT_WEBHOOK_URL ?? '',
 };
 
-export default config;
+export default environment;
