+++ packages/functions/src/config/index.ts
@@ -1,3 +1,3 @@
 export { default as supabase } from './supabase';
-export { default as config } from './variables';
+export { default as environment } from './environment';
 export { default as SupportedChains } from './chains';
